requirements.txt
torch>=2.0.0
pytorch-lightning>=2.1.0
torchmetrics>=1.2.0
pandas>=2.1.0
PyYAML>=6.0
numpy>=1.24.0
scipy>=1.10.0